// crm-client/src/types/types.ts
// export type Client = {
//   id: string;
//   name: string;
//   phone: string;
//   email: string;
//   status: string;
//   createdAt: string;
//   updatedAt: string;
// };

export interface Client {
  id: string;
  name: string;
  phone: string;
  email: string;
  source: string;
  status: string;
  createdAt: string;
  updatedAt: string;
}